stand by

+++ cli.js
@@ -19,18 +19,18 @@ const fs = require("fs");
 //                          global functions                          //
 ////////////////////////////////////////////////////////////////////////
 
-async function checkIfConfigurationFolderIsCreated() {
+async function checkIfConfigurationDirIsCreated() {
   try {
     const isExist = await fs.accessSync(
       `${homedir}/.vimConfiguration`,
       fs.constants.F_OK
     );
   } catch (e) {
-    await createConfigurationFolder();
+    await createConfigurationDir();
   }
 }
 
-async function createConfigurationFolder() {
+async function createConfigurationDir() {
   try {
     await fs.mkdirSync(`${homedir}/.vimConfiguration`);
     console.log("dossier de configuration crée");
@@ -39,34 +39,83 @@ async function createConfigurationFolder() {
   }
 }
 
-function checkArguments() {
-  if (args[0]) {
-    return args[1] !== undefined ? args[0] : false;
-  } else {
-    return false;
-  }
+function checkNameExist() {
+  return args[1];
 }
+
+// async function checkIfBackupDirIsCreated() {
+//   try {
+//     const isExist = await fs.accessSync(
+//       `${homedir}/.vimConfiguration/oldConfig`,
+//       fs.constants.F_OK
+//     );
+//   } catch (e) {
+//     await createBackupDir();
+//   }
+// }
+
+// async function createBackupDir() {
+//   try {
+//     await fs.mkdirSync(`${homedir}/.vimConfiguration/oldConfig`);
+//     console.log("dossier de backup crée");
+//   } catch (e) {
+//     console.log(e);
+//   }
+// }
+
+// async function updateBackupDir() {
+//   try {
+//     const pluggedDirExist = await fs.accessSync(
+//       `${homedir}/.vimConfiguration/oldConfig/plugged`,
+//       fs.constants.F_OK
+//     );
+//     const vimrcExist = await fs.accessSync(
+//       `${homedir}/.vimConfiguration/oldConfig/.vimrc`,
+//       fs.constants.F_OK
+//     );
+
+//     console.log(pluggedDirExist);
+//     console.log(vimrcExist);
+
+//     if (pluggedDirExist) {
+//       await fs.rmdirSync(`${homedir}/.vimConfiguration/oldConfig/plugged`);
+//     }
+//     if (vimrcExist) {
+//       await fs.unlinkSync(`${homedir}/.vimConfiguration/oldConfig/.vimrc`);
+//     }
+
+//     await fs.copyFileSync(
+//       `${homedir}/.vimrc`,
+//       `${homedir}/.vimConfiguration/oldConfig/.vimrc`
+//     );
+
+//     // boucle pour copier tout ce qu'il ya dans .vim/plugged
+//   } catch (e) {
+//     console.log(e);
+//   }
+// }
 
 ////////////////////////////////////////////////////////////////////////
 //                           create section                           //
 ////////////////////////////////////////////////////////////////////////
 
 async function createVimConfiguation() {
-  await checkIfConfigurationFolderIsCreated();
+  if (checkNameExist()) {
+    await checkIfConfigurationDirIsCreated();
 
-  try {
-    await fs.mkdirSync(`${homedir}/.vimConfiguration/${args[1]}/plugged`, {
-      recursive: true
-    });
-    await fs.writeFileSync(
-      `${homedir}/.vimConfiguration/${args[1]}/.vimrc`,
-      ""
-    );
-  } catch (e) {
-    console.log(e);
-  }
+    try {
+      await fs.mkdirSync(`${homedir}/.vimConfiguration/${args[1]}/plugged`, {
+        recursive: true
+      });
+      await fs.writeFileSync(
+        `${homedir}/.vimConfiguration/${args[1]}/.vimrc`,
+        ""
+      );
+    } catch (e) {
+      console.log(e);
+    }
 
-  console.log(`
+    console.log(`
     new configuration created with name: ${args[1]}
 
     now you can:
@@ -74,6 +123,13 @@ async function createVimConfiguation() {
       - remove her
 
     `);
+  } else {
+    console.log(`
+
+      Veuillez ajouter un nom à la configuration
+
+      `);
+  }
 }
 
 ////////////////////////////////////////////////////////////////////////
@@ -81,8 +137,12 @@ async function createVimConfiguation() {
 ////////////////////////////////////////////////////////////////////////
 
 async function switchVimConfiguation() {
-  await checkIfConfigurationFolderIsCreated();
-  console.log("switched");
+  await checkIfConfigurationDirIsCreated();
+  try {
+    console.log("switch");
+  } catch (e) {
+    console.log(e);
+  }
 }
 
 ////////////////////////////////////////////////////////////////////////
@@ -90,14 +150,38 @@ async function switchVimConfiguation() {
 ////////////////////////////////////////////////////////////////////////
 
 async function removeVimConfiguation() {
-  await checkIfConfigurationFolderIsCreated();
+  await checkIfConfigurationDirIsCreated();
   try {
     await fs.rmdirSync(`${homedir}/.vimConfiguration/${args[1]}/plugged`);
     await fs.unlinkSync(`${homedir}/.vimConfiguration/${args[1]}/.vimrc`);
     await fs.rmdirSync(`${homedir}/.vimConfiguration/${args[1]}`);
-    console.log(`La configuration ${args[1]} a bien été supprimée`);
+    console.log(`
+
+      La configuration ${args[1]} a bien été supprimée
+
+      `);
   } catch (e) {
     console.log("La configuration que vous voulez effacer n'existe pas");
+  }
+}
+
+////////////////////////////////////////////////////////////////////////
+//                            list section                            //
+////////////////////////////////////////////////////////////////////////
+
+async function listVimConfiguration() {
+  await checkIfConfigurationDirIsCreated();
+  try {
+    const allConfig = await fs.readdirSync(`${homedir}/.vimConfiguration`);
+    console.log(`
+    Configurations actuel:
+    `);
+    allConfig.forEach(element => {
+      console.log(`    - ${element} `);
+    });
+    console.log(` `);
+  } catch (e) {
+    console.log("Vos configurations n'ont pas pu être listé");
   }
 }
 
@@ -110,6 +194,7 @@ const helpString = `
 
       Commands: 
         create <name> - (create new configuration named $name)
+        list - (list all configurations)
         switch <name> - (switch current configuration with $name)
         remove <name> - (remove $name configuration)
 
@@ -126,7 +211,7 @@ const helpString = `
 
 const [, , ...args] = process.argv;
 
-switch (checkArguments()) {
+switch (args[0]) {
   case "create":
     createVimConfiguation();
     break;
@@ -136,8 +221,11 @@ switch (checkArguments()) {
   case "remove":
     removeVimConfiguation();
     break;
+  case "list":
+    listVimConfiguration();
+    break;
   case "-h":
   case "--help":
-  default:
     console.log(helpString);
+    break;
 }
